+++ projects/Mallard/src/screens/settings-screen.tsx
@@ -1,156 +1,34 @@
-import React, { useContext, useState } from 'react'
-import { Text, Alert, StyleSheet } from 'react-native'
 import AsyncStorage from '@react-native-community/async-storage'
-
-import { List } from 'src/components/lists/list'
-import { withNavigation, NavigationInjectedProps } from 'react-navigation'
-import { useSettings } from 'src/hooks/use-settings'
-import { UiBodyCopy } from 'src/components/styled-text'
-import { APP_DISPLAY_NAME, FEEDBACK_EMAIL } from 'src/helpers/words'
-import { clearCache } from 'src/helpers/fetch/cache'
-import { Heading, Footer } from 'src/components/layout/ui/row'
-import { getVersionInfo } from 'src/helpers/settings'
-import { metrics } from 'src/theme/spacing'
-import { ScrollContainer } from 'src/components/layout/ui/container'
-import { WithAppAppearance } from 'src/theme/appearance'
+import React, { useContext, useState } from 'react'
+import { Alert, StyleSheet, Text } from 'react-native'
+import { NavigationInjectedProps } from 'react-navigation'
 import {
-    useIdentity,
     AuthContext,
     useAuth,
+    useIdentity,
 } from 'src/authentication/auth-context'
 import { RightChevron } from 'src/components/icons/RightChevron'
-import { getFont } from 'src/theme/typography'
-import { color } from 'src/theme/color'
+import { ScrollContainer } from 'src/components/layout/ui/container'
+import { Footer, Heading } from 'src/components/layout/ui/row'
+import { List } from 'src/components/lists/list'
+import { UiBodyCopy } from 'src/components/styled-text'
+import { clearCache } from 'src/helpers/fetch/cache'
+import { getVersionInfo } from 'src/helpers/settings'
+import { APP_DISPLAY_NAME, FEEDBACK_EMAIL } from 'src/helpers/words'
+import { useSettings } from 'src/hooks/use-settings'
 import { routeNames } from 'src/navigation/routes'
-
-const DevZone = withNavigation(({ navigation }: NavigationInjectedProps) => {
-    const [settings, setSetting] = useSettings()
-    const { status } = useContext(AuthContext)
-    const { apiUrl } = settings
-    return (
-        <>
-            <Heading>🦆 SECRET DUCK MENU 🦆</Heading>
-            <Footer>
-                <UiBodyCopy>
-                    Only wander here if you know what you are doing!!
-                </UiBodyCopy>
-            </Footer>
-            <List
-                onPress={({ onPress }) => onPress()}
-                data={[
-                    {
-                        key: 'Downloads',
-                        title: 'Manage issues',
-                        data: {
-                            onPress: () => {
-                                navigation.navigate(routeNames.Downloads)
-                            },
-                        },
-                    },
-                    {
-                        key: 'Endpoints',
-                        title: 'API Endpoint',
-                        explainer: apiUrl,
-                        data: {
-                            onPress: () => {
-                                navigation.navigate(routeNames.Endpoints)
-                            },
-                        },
-                    },
-                    {
-                        key: 'Clear caches',
-                        title: 'Clear caches',
-                        data: {
-                            onPress: () => {
-                                Alert.alert(
-                                    'Clear caches',
-                                    'You sure?',
-                                    [
-                                        {
-                                            text: 'Delete fetch cache',
-                                            onPress: () => {
-                                                clearCache()
-                                            },
-                                        },
-                                        {
-                                            text: 'Delete EVERYTHING',
-                                            onPress: () => {
-                                                AsyncStorage.clear()
-                                            },
-                                        },
-                                        {
-                                            style: 'cancel',
-                                            text: `No don't do it`,
-                                        },
-                                    ],
-                                    { cancelable: false },
-                                )
-                            },
-                        },
-                    },
-                    {
-                        key: 'Re-start onboarding',
-                        title: 'Re-start onboarding',
-                        data: {
-                            onPress: () => {
-                                // go back to the main to simulate a fresh app
-                                setSetting('hasOnboarded', false)
-                                navigation.navigate('Onboarding')
-                            },
-                        },
-                    },
-                    {
-                        key: 'Hide this menu',
-                        title: 'Hide this menu',
-                        explainer:
-                            'Tap the version 7 times & confirm the fake scary message to bring it back',
-                        data: {
-                            onPress: () => {
-                                setSetting('isUsingProdDevtools', false)
-                            },
-                        },
-                    },
-                    {
-                        key: 'Build id',
-                        title: 'Build',
-                        explainer: getVersionInfo().commitId,
-                        data: {
-                            onPress: () => {},
-                        },
-                    },
-                ]}
-            />
-            <Heading>Your settings</Heading>
-            <List
-                onPress={() => {}}
-                data={Object.entries(settings)
-                    .map(([title, explainer]) => ({
-                        key: title,
-                        title,
-                        explainer: explainer + '',
-                    }))
-                    .concat([
-                        {
-                            key: 'Authentication details',
-                            title: 'Authentication details',
-                            explainer: `Signed in status ${status.type} : ${
-                                status.type === 'authed'
-                                    ? status.data.type
-                                    : '_'
-                            }`,
-                        },
-                    ])}
-            />
-        </>
-    )
-})
+import { WithAppAppearance } from 'src/theme/appearance'
+import { color } from 'src/theme/color'
+import { metrics } from 'src/theme/spacing'
+import { getFont } from 'src/theme/typography'
+import { DevZone } from './settings/dev-zone'
 
 const SettingsScreen = ({ navigation }: NavigationInjectedProps) => {
     const [settings, setSetting] = useSettings()
     const { isUsingProdDevtools } = settings
     const signInHandler = useIdentity()
     const authHandler = useAuth()
-    const [versionClickedTimes, setVersionClickedTimes] = useState(0)
+    const [, setVersionClickedTimes] = useState(0)
     const { signOut } = useContext(AuthContext)
     const styles = StyleSheet.create({
         signOut: {
